migrate: move vetCenterLocationListing to products directory
- Migrated vetCenterLocationListing product files to src/products/vetCenterLocationListing/
- Moved and renamed index.tsx -> template.tsx
- Moved and renamed index.test.tsx -> template.test.tsx
- Moved VetCenterLocationInfo component and test files
- Moved query.ts from src/data/queries/ to src/products/vetCenterLocationListing/
- Moved test file and renamed vetCenterLocationListing.test.tsx -> query.test.ts
- Moved formatted type from src/types/formatted/ to src/products/vetCenterLocationListing/formatted-type.ts
- Moved mock file from src/mocks/ to src/products/vetCenterLocationListing/mock.json
- Updated all import statements across the codebase
- Fixed import paths in migrated files
- All TypeScript errors resolved
- All tests passing (7 query tests, 12 template tests, 17 component tests)
- Updated migration guide with completed status
- Added CODEOWNERS entry for new product directory
- Verified dev server runs and vetCenterLocationListing pages load correctly
