+++ src/data/queries/index.ts
@@ -50,7 +50,7 @@ import * as VamcSystemVaPolice from '@/products/vamcSystemVaPolice/query'
 import * as VetCenter from '@/products/vetCenter/query'
 import * as VetCenterHealthServices from '@/data/queries/vetCenterHealthServices'
 import * as VetCenterOutstation from '@/data/queries/vetCenterOutstation'
-import * as VetCenterLocationListing from '@/data/queries/vetCenterLocationListing'
+import * as VetCenterLocationListing from '@/products/vetCenterLocationListing/query'
 import * as Wysiwyg from '@/data/queries/wysiwyg'
 import * as LeadershipListing from '@/products/leadershipListing/query'
 import * as VbaFacility from '@/data/queries/vbaFacility'
