+++ src/products/vetCenterLocationListing/template.test.tsx
@@ -1,8 +1,8 @@
 import React from 'react'
 import { render, screen } from '@testing-library/react'
-import { VetCenterLocationListing } from './index'
-import { formatter } from '@/data/queries/vetCenterLocationListing'
-import drupalMockData from '@/mocks/vetCenterLocationListing.mock.json'
+import { VetCenterLocationListing } from './template'
+import { formatter } from '@/products/vetCenterLocationListing/query'
+import drupalMockData from '@/products/vetCenterLocationListing/mock.json'
 import mockCap from '@/mocks/vetCenterCap.mock.json'
 import mockOutstation from '@/mocks/vetCenterOutstation.mock.json'
 import { NodeVetCenterCap } from '@/types/drupal/node'
